Food Saver Various Wireframes
This is a screenshot of the wireframes for the various android app UX/UI layouts. These are mockups of how we can create the front-end for the application. Each wireframe contains a menu widget to go through the various layouts of the application. There is also a search feature widget if you are looking for a specific food item it should pull up a list of all the locations containing that item. This mock up is based on using Yelp's API and a Google Maps API.